# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WIT Sprint OS API
  version: "1.0.0"
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    REST API for the WIT Sprint OS delivery hierarchy.

    ## Naming

    The URL vocabulary predates the UI's, and the two do not match. Read it as:

    | URL / payload | The UI calls it |
    | ------------- | --------------- |
    | `product`     | **Module**      |
    | `module`      | **Component**   |

    The hierarchy is `client → project → product → module → sprint → task`,
    which the UI presents as Client → Project → Module → Component → Sprint.

    ## Conventions

    - **PATCH is partial.** Only the keys present in the body are written; an
      absent key leaves its column untouched. This is what makes two concurrent
      edits to different fields both survive. `PATCH /products/{id}/current-sprint`
      is the single exception where an explicit `null` differs from an absent
      key — see that operation.
    - **Ids may be supplied or generated.** Send `id` on create to choose it;
      omit it and the server mints one. A duplicate answers `409`.
    - **Lineage is enforced.** A child may only reference a parent inside its own
      product: a sprint's component, a task's backlog item, a report's sprint. A
      mismatch is a `400`, not a silent link.
    - **Timestamps** are RFC 3339 with an offset (`2026-07-23T02:35:08.59+07:00`).
...
      RFC 3339.
...

    ## Authentication

...
    it stands.

...
security: []
servers:
...
paths:
...
  /sprints/{sprintId}:
    parameters:
      - $ref: '#/components/parameters/SprintId'
    get:
      tags: [sprints]
      summary: Get a sprint
      operationId: getSprint
      responses:
        '200':
          description: The sprint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [sprints]
      summary: Update a sprint
      operationId: updateSprint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SprintUpdate' }
      responses:
        '200':
          description: The updated sprint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [sprints]
      summary: Delete a sprint
      description: Cascades to its tasks, members and committed backlog links.
      operationId: deleteSprint
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /sprints/{sprintId}/tasks:
    parameters:
      - $ref: '#/components/parameters/SprintId'
    get:
      tags: [tasks]
      summary: List a sprint's tasks
      operationId: listTasksBySprint
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of tasks.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [tasks]
      summary: Create a task
      description: '`backlogItemId` must belong to the same product as the sprint.'
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreate' }
      responses:
        '201':
          description: The created task.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400':
          description: The backlog item does not exist or belongs to a different module.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [tasks]
      summary: Get a task
      operationId: getTask
      responses:
        '200':
          description: The task.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tasks]
      summary: Update a task
      description: |
        Partial, which is what lets someone rename a card while a colleague
        drags it to another column without either edit being lost.
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdate' }
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [tasks]
      summary: Delete a task
      description: Cascades to its DoD checklist.
      operationId: deleteTask
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /tasks/{taskId}/column:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    patch:
      tags: [tasks]
      summary: Move a task to another board column
      operationId: moveTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boardColumn]
              properties:
                boardColumn: { $ref: '#/components/schemas/BoardColumn' }
      responses:
        '200':
          description: The moved task.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/dod:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [tasks]
      summary: List a task's Definition of Done
      description: |
        **Not paginated, and not bounded.** `?limit` is ignored and the whole set
        is returned. A load test found one sprint holding 4,898 committed items
        answering 1.3 MB to a request that asked for 5, so treat the size of this
        response as a function of your data, not of the request. Paginating these
        is a known outstanding fix.

        Ordered by `position`.
      operationId: listTaskDod
      responses:
        '200':
          description: The checklist.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskDodItem' }
        '503': { $ref: '#/components/responses/Unavailable' }
    put:
      tags: [tasks]
      summary: Add or replace one checklist entry
      description: Keyed on `position` — repeating a position overwrites it.
      operationId: setTaskDodItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position, label]
              properties:
                position: { type: integer, minimum: 0 }
                label: { type: string }
                done: { type: boolean, default: false }
      responses:
        '200':
          description: The entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskDodItem' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /tasks/{taskId}/dod/{position}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
      - name: position
        in: path
        required: true
        schema: { type: integer, minimum: 0 }
    patch:
      tags: [tasks]
      summary: Toggle one checklist entry
      operationId: toggleTaskDodItem
      responses:
        '200':
          description: The toggled entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskDodItem' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------- members ---
  /members:
    get:
      tags: [members]
      summary: List members
      operationId: listMembers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of members.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [members]
      summary: Create a member
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberCreate' }
      responses:
        '201':
          description: The created member.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: That email is already taken.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [members]
      summary: Get a member
      operationId: getMember
      responses:
        '200':
          description: The member.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [members]
      summary: Update a member
      operationId: updateMember
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberUpdate' }
      responses:
        '200':
          description: The updated member.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [members]
      summary: Delete a member
      description: |
        Their sprint memberships go with them; tasks they were assigned survive
        with `assigneeId` set to null.
      operationId: deleteMember
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ----------------------------------------------------------- decisions ---
  /products/{productId}/decisions:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [decisions]
      summary: List a product's decision log
      operationId: listDecisionsByProduct
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of decisions, most recent first.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Decision' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [decisions]
      summary: Record a decision
      operationId: createDecision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecisionCreate' }
      responses:
        '201':
          description: The created decision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Decision' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /decisions/{decisionId}:
    parameters:
      - name: decisionId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [decisions]
      summary: Update a decision
      operationId: updateDecision
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecisionUpdate' }
      responses:
        '200':
          description: The updated decision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Decision' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [decisions]
      summary: Delete a decision
      operationId: deleteDecision
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ------------------------------------------------------------- reports ---
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Clamped to 1000; anything unparseable falls back to the default.
      schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
    Offset:
      name: offset
      in: query
      description: |
        Rows to skip. Postgres walks and discards them, so deep offsets get
        progressively more expensive — prefer narrowing by parent id. Values
        above 1,000,000 are refused with a 400 rather than silently answered
        with a different page.
      schema: { type: integer, minimum: 0, maximum: 1000000, default: 0 }
...
    ProductId:
      name: productId
      in: path
      required: true
      description: 'A product id — the UI calls this a Module.'
      schema: { type: string }
...
    SprintId: { name: sprintId, in: path, required: true, schema: { type: string } }
    TaskId:   { name: taskId,   in: path, required: true, schema: { type: string } }
    MemberId: { name: memberId, in: path, required: true, schema: { type: string } }
...
  headers:
    XHasMore:
      description: Present and `true` when another page exists.
      schema: { type: string, enum: ['true'] }
  responses:
    BadRequest:
      description: The request is malformed, or a value is out of range.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such record.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        A uniqueness clash, or a row was busy and the write should be retried.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: The body is over 1 MB.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Retryable. Either the request took too long — the statement timeout is
        5 s and the whole request is capped at 15 s — or the database could not
        take it: out of connection slots, restarting, or the connection dropped.
        Never a bug on the caller's side, and always worth retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, examples: ['name is required'] }

    # ------------------------------------------------------------- enums ---
...
    Risk:              { type: string, enum: [low, medium, high] }
...
    BoardColumn:
      type: string
      enum: [selected, ready, in-progress, in-review, qa, done, blocked]
...
    Sprint:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        moduleId: { type: [string, 'null'], description: The component that owns this sprint. }
        number: { type: integer, minimum: 1, maximum: 100000 }
        name: { type: string }
        goal: { type: string }
        startDate: { $ref: '#/components/schemas/DateOnly' }
        endDate: { $ref: '#/components/schemas/DateOnly' }
        workingDays: { type: integer, minimum: 0 }
        daysLeft: { type: integer, minimum: 0 }
        status: { $ref: '#/components/schemas/SprintStatus' }
        committed: { type: integer, minimum: 0 }
        completed: { type: integer, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    SprintUpdate:
      type: object
      properties:
        moduleId: { type: string }
        number: { type: integer, minimum: 1, maximum: 100000 }
        name: { type: string }
        goal: { type: string }
        startDate: { $ref: '#/components/schemas/DateOnly' }
        endDate: { $ref: '#/components/schemas/DateOnly' }
        workingDays: { type: integer, minimum: 0 }
        daysLeft: { type: integer, minimum: 0 }
        status: { $ref: '#/components/schemas/SprintStatus' }
        committed: { type: integer, minimum: 0 }
        completed: { type: integer, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
...
    Task:
      type: object
      properties:
        id: { type: string }
        sprintId: { type: string }
        backlogItemId: { type: string }
        title: { type: string }
        moduleName: { type: string, description: Denormalised label shown on the card. }
        assigneeId: { type: [string, 'null'] }
        estimate: { type: integer, minimum: 0 }
        boardColumn: { $ref: '#/components/schemas/BoardColumn' }
        priority: { type: string, description: 'Free text; the UI offers low, medium, high.' }
        blockedReason: { type: [string, 'null'] }
        blockedDays: { type: [integer, 'null'], minimum: 0 }
        offGoal: { type: boolean }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    TaskCreate:
      type: object
      required: [title, backlogItemId]
      properties:
        id: { type: string }
        backlogItemId: { type: string, description: Must belong to the sprint's product. }
        title: { type: string }
        moduleName: { type: string }
        assigneeId: { type: string }
        estimate: { type: integer, minimum: 0 }
        boardColumn: { $ref: '#/components/schemas/BoardColumn' }
        priority: { type: string, default: medium }
        blockedReason: { type: string }
        blockedDays: { type: integer, minimum: 0 }
        offGoal: { type: boolean, default: false }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        moduleName: { type: string }
        assigneeId: { type: string }
        estimate: { type: integer, minimum: 0 }
        boardColumn: { $ref: '#/components/schemas/BoardColumn' }
        priority: { type: string }
        blockedReason: { type: string }
        blockedDays: { type: integer, minimum: 0 }
        offGoal: { type: boolean }
    TaskDodItem:
      type: object
      properties:
        taskId: { type: string }
        position: { type: integer, minimum: 0 }
        label: { type: string }
        done: { type: boolean }

    # ----------------------------------------------------------- members ---
...